+++ HISTORY.md
@@ -34,9 +34,10 @@
 * objects: enabled custom categories for the stage
 * store: load user defined palettes
 * gui: restore custom palettes when activating a scene
-* added new category feature to palette context menu
+* objects: added new category feature to palette context menu
 * gui, objects, scenes: fixed costume maximum extent
 * gui: cleaned up hidden menu
+* objects, gui: added delete category feature to palette context menu
 
 ### 2021-07-21
 * user defined custom block palettes, under construction